2.5.0 release

+++ avt/DBAtts/MetaData/avtSimulationInformation.C
@@ -1,6 +1,6 @@
 /*****************************************************************************
 *
-* Copyright (c) 2000 - 2011, Lawrence Livermore National Security, LLC
+* Copyright (c) 2000 - 2012, Lawrence Livermore National Security, LLC
 * Produced at the Lawrence Livermore National Laboratory
 * LLNL-CODE-442911
 * All rights reserved.
@@ -816,7 +816,7 @@ avtSimulationInformation::RemoveGenericCommands(int index)
 int
 avtSimulationInformation::GetNumGenericCommands() const
 {
-    return genericCommands.size();
+    return static_cast<int>(genericCommands.size());
 }
 
 // ****************************************************************************
@@ -967,7 +967,7 @@ avtSimulationInformation::RemoveCustomCommands(int index)
 int
 avtSimulationInformation::GetNumCustomCommands() const
 {
-    return customCommands.size();
+    return static_cast<int>(customCommands.size());
 }
 
 // ****************************************************************************
