+++ avt/DBAtts/SIL/avtSILEnumeratedNamespace.C
@@ -1,6 +1,6 @@
 /*****************************************************************************
 *
-* Copyright (c) 2000 - 2011, Lawrence Livermore National Security, LLC
+* Copyright (c) 2000 - 2012, Lawrence Livermore National Security, LLC
 * Produced at the Lawrence Livermore National Laboratory
 * LLNL-CODE-442911
 * All rights reserved.
@@ -133,7 +133,7 @@ avtSILEnumeratedNamespace::ContainsElement(int e) const
     }
     else if (sortedElems)
     {
-        int min = 0, max = elements.size()-1;
+        int min = 0, max = static_cast<int>(elements.size())-1;
         int mid = (max+min)/2;
         while (min <= max)
         {
