+++ avt/DBAtts/SIL/avtSILRestrictionTraverser.C
@@ -1,6 +1,6 @@
 /*****************************************************************************
 *
-* Copyright (c) 2000 - 2011, Lawrence Livermore National Security, LLC
+* Copyright (c) 2000 - 2012, Lawrence Livermore National Security, LLC
 * Produced at the Lawrence Livermore National Laboratory
 * LLNL-CODE-442911
 * All rights reserved.
@@ -1086,7 +1086,7 @@ avtSILRestrictionTraverser::AddMaterialList(int chunk, MaterialList &l,
     {
         materialListForChunk.push_back(-1);
     }
-    materialListForChunk[chunk] = materialList.size();
+    materialListForChunk[chunk] = static_cast<int>(materialList.size());
     materialList.push_back(l);
     shouldMatSelect.push_back(shouldMatSel);
 }
