+++ avt/Expressions/Management/avtExprNode.C
@@ -1,6 +1,6 @@
 /*****************************************************************************
 *
-* Copyright (c) 2000 - 2011, Lawrence Livermore National Security, LLC
+* Copyright (c) 2000 - 2012, Lawrence Livermore National Security, LLC
 * Produced at the Lawrence Livermore National Laboratory
 * LLNL-CODE-442911
 * All rights reserved.
@@ -789,7 +789,7 @@ avtFunctionExpr::CreateFilters(ExprPipelineState *state)
 
     // Take the stack of variable names and feed them to the function in
     // reverse order.
-    int size = inputStack.size();
+    int size = static_cast<int>(inputStack.size());
     for (i = 0; i < size; i++)
     {
         string inputName = inputStack.back();
