+++ avt/Expressions/Math/avtBase10LogExpression.C
@@ -1,6 +1,6 @@
 /*****************************************************************************
 *
-* Copyright (c) 2000 - 2011, Lawrence Livermore National Security, LLC
+* Copyright (c) 2000 - 2012, Lawrence Livermore National Security, LLC
 * Produced at the Lawrence Livermore National Laboratory
 * LLNL-CODE-442911
 * All rights reserved.
@@ -124,13 +124,13 @@ avtBase10LogExpression::~avtBase10LogExpression()
 
 void
 avtBase10LogExpression::DoOperation(vtkDataArray *in, vtkDataArray *out,
-                                int ncomponents, int ntuples)
+                                    int ncomponents, int ntuples)
 {
     for (int i = 0 ; i < ntuples ; i++)
     {
         for (int j = 0 ; j < ncomponents ; j++)
         {
-            float f = in->GetComponent(i, j);
+            double f = in->GetComponent(i, j);
             if (f <= 0)
             {
                 if (useDefaultOnError)
