+++ avt/Pipeline/Sinks/avtTerminatingSink.C
@@ -1,6 +1,6 @@
 /*****************************************************************************
 *
-* Copyright (c) 2000 - 2011, Lawrence Livermore National Security, LLC
+* Copyright (c) 2000 - 2012, Lawrence Livermore National Security, LLC
 * Produced at the Lawrence Livermore National Laboratory
 * LLNL-CODE-442911
 * All rights reserved.
@@ -210,7 +210,7 @@ avtTerminatingSink::Execute(avtContract_p contract)
 
             if (DebugStream::Level3())
             {
-                unsigned int size = 0, rss = 0;
+                unsigned long size = 0, rss = 0;
                 GetMemorySize(size, rss);
                 if (size > 0 && rss > 0)
                 {
@@ -242,7 +242,7 @@ avtTerminatingSink::Execute(avtContract_p contract)
 
                 if (DebugStream::Level3())
                 {
-                    unsigned int size = 0, rss = 0;
+                    unsigned long size = 0, rss = 0;
                     GetMemorySize(size, rss);
                     if (size > 0 && rss > 0)
                     {
